+++ frontend/vue/components/Stations/Queue.vue
@@ -81,7 +81,9 @@ export default {
     },
     methods: {
         formatTime(time) {
-            return this.getDateTime(time).toLocaleString(DateTime.TIME_WITH_SECONDS);
+            return this.getDateTime(time).toLocaleString(
+                {...DateTime.TIME_WITH_SECONDS, ...App.time_config}
+            );
         },
         formatRelativeTime(time) {
             return this.getDateTime(time).toRelative();